SOS-1532
