+++ portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/addusersoadditionalemailaddress/AddUserSOAdditionalEmailAddressTests.html
@@ -12,6 +12,7 @@
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/selectregularrolessouser/SelectRegularRolesSOUserTest.html">SelectRegularRolesSOUserTest</a></td></tr>
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/editsouserpassword/EditSOUserPasswordTest.html">EditSOUserPasswordTest</a></td></tr>
 <tr><td><a href="AddUserSOAdditionalEmailAddressTest.html">AddUserSOAdditionalEmailAddressTest</a></td></tr>
+<tr><td><a href="ViewProfileAdditionalEmailAddressTest.html">ViewProfileAdditionalEmailAddressTest</a></td></tr>
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/addsouser/TearDownSOUserTest.html">TearDownSOUserTest</a></td></tr>
 </tbody></table>
 </body>
