+++ portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/addusersoaddress/AddUserSOAddressTests.html
@@ -12,6 +12,7 @@
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/selectregularrolessouser/SelectRegularRolesSOUserTest.html">SelectRegularRolesSOUserTest</a></td></tr>
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/editsouserpassword/EditSOUserPasswordTest.html">EditSOUserPasswordTest</a></td></tr>
 <tr><td><a href="AddUserSOAddressTest.html">AddUserSOAddressTest</a></td></tr>
+<tr><td><a href="ViewProfileAddressTest.html">ViewProfileAddressTest</a></td></tr>
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/addsouser/TearDownSOUserTest.html">TearDownSOUserTest</a></td></tr>
 </tbody></table>
 </body>
