+++ portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/addusersoexpertise/AddUserSOExpertiseTests.html
@@ -12,6 +12,7 @@
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/selectregularrolessouser/SelectRegularRolesSOUserTest.html">SelectRegularRolesSOUserTest</a></td></tr>
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/editsouserpassword/EditSOUserPasswordTest.html">EditSOUserPasswordTest</a></td></tr>
 <tr><td><a href="AddUserSOExpertiseTest.html">AddUserSOExpertiseTest</a></td></tr>
+<tr><td><a href="ViewProfileExpertiseTest.html">ViewProfileExpertiseTest</a></td></tr>
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/addsouser/TearDownSOUserTest.html">TearDownSOUserTest</a></td></tr>
 </tbody></table>
 </body>
