+++ portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/addusersoinstantmessenger/ViewProfileInstantMessengerTest.html
