+++ portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/addusersophonenumber/AddUserSOPhoneNumberTests.java
@@ -33,6 +33,7 @@ public class AddUserSOPhoneNumberTests extends BaseTestSuite {
 		testSuite.addTestSuite(SelectRegularRolesSOUserTest.class);
 		testSuite.addTestSuite(EditSOUserPasswordTest.class);
 		testSuite.addTestSuite(AddUserSOPhoneNumberTest.class);
+		testSuite.addTestSuite(ViewProfilePhoneNumberTest.class);
 		testSuite.addTestSuite(TearDownSOUserTest.class);
 
 		return testSuite;
