+++ portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/addusersosms/AddUserSOSMSTests.html
@@ -12,6 +12,7 @@
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/selectregularrolessouser/SelectRegularRolesSOUserTest.html">SelectRegularRolesSOUserTest</a></td></tr>
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/editsouserpassword/EditSOUserPasswordTest.html">EditSOUserPasswordTest</a></td></tr>
 <tr><td><a href="AddUserSOSMSTest.html">AddUserSOSMSTest</a></td></tr>
+<tr><td><a href="ViewProfileSMSTest.html">ViewProfileSMSTest</a></td></tr>
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/addsouser/TearDownSOUserTest.html">TearDownSOUserTest</a></td></tr>
 </tbody></table>
 </body>
