+++ portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/addusersosms/AddUserSOSMSTests.java
@@ -33,6 +33,7 @@ public class AddUserSOSMSTests extends BaseTestSuite {
 		testSuite.addTestSuite(SelectRegularRolesSOUserTest.class);
 		testSuite.addTestSuite(EditSOUserPasswordTest.class);
 		testSuite.addTestSuite(AddUserSOSMSTest.class);
+		testSuite.addTestSuite(ViewProfileSMSTest.class);
 		testSuite.addTestSuite(TearDownSOUserTest.class);
 
 		return testSuite;
