+++ portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/addusersosms/ViewProfileSMSTest.html
