+++ portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/addusersosocialnetwork/ViewProfileSocialNetworkTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.socialofficeprofile.profile.viewprofileadditionalemailaddress;
+package com.liferay.portalweb.socialofficeprofile.profile.addusersosocialnetwork;
 
 import com.liferay.portalweb.portal.BaseTestCase;
 import com.liferay.portalweb.portal.util.RuntimeVariables;
@@ -20,9 +20,8 @@ import com.liferay.portalweb.portal.util.RuntimeVariables;
 /**
  * @author Brian Wing Shun Chan
  */
-public class ViewProfileAdditionalEmailAddressTest extends BaseTestCase {
-	public void testViewProfileAdditionalEmailAddress()
-		throws Exception {
+public class ViewProfileSocialNetworkTest extends BaseTestCase {
+	public void testViewProfileSocialNetwork() throws Exception {
 		selenium.selectWindow("null");
 		selenium.selectFrame("relative=top");
 		selenium.open("/web/socialoffice01/so/profile");
@@ -34,15 +33,27 @@ public class ViewProfileAdditionalEmailAddressTest extends BaseTestCase {
 		assertEquals(RuntimeVariables.replace("socialoffice01@liferay.com"),
 			selenium.getText("//div[@class='lfr-contact-extra']"));
 		assertTrue(selenium.isVisible(
-				"//div[@class='section field-group lfr-user-email-addresses']/h3"));
-		assertEquals(RuntimeVariables.replace("Additional Email Addresses:"),
+				"//div[@class='section field-group lfr-user-social-network']/h3"));
+		assertEquals(RuntimeVariables.replace("Social Network:"),
 			selenium.getText(
-				"//div[@class='section field-group lfr-user-email-addresses']/h3"));
-		assertEquals(RuntimeVariables.replace("Email Address"),
+				"//div[@class='section field-group lfr-user-social-network']/h3"));
+		assertEquals(RuntimeVariables.replace("Facebook"),
 			selenium.getText(
-				"//div[@class='section field-group lfr-user-email-addresses']/ul/li/span"));
-		assertEquals(RuntimeVariables.replace("socialoffice02@liferay.com"),
+				"//div[@class='section field-group lfr-user-social-network']/ul/li/span"));
+		assertEquals(RuntimeVariables.replace("socialoffice01"),
 			selenium.getText(
-				"//div[@class='section field-group lfr-user-email-addresses']/ul/li/span[2]/a"));
+				"//div[@class='section field-group lfr-user-social-network']/ul/li/span[2]"));
+		assertEquals(RuntimeVariables.replace("MySpace"),
+			selenium.getText(
+				"//div[@class='section field-group lfr-user-social-network']/ul/li[2]/span"));
+		assertEquals(RuntimeVariables.replace("socialoffice01"),
+			selenium.getText(
+				"//div[@class='section field-group lfr-user-social-network']/ul/li[2]/span[2]"));
+		assertEquals(RuntimeVariables.replace("Twitter"),
+			selenium.getText(
+				"//div[@class='section field-group lfr-user-social-network']/ul/li[3]/span"));
+		assertEquals(RuntimeVariables.replace("socialoffice01"),
+			selenium.getText(
+				"//div[@class='section field-group lfr-user-social-network']/ul/li[3]/span[2]"));
 	}
 }
