+++ portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/addusersowebsite/AddUserSOWebsiteTests.html
@@ -12,6 +12,7 @@
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/selectregularrolessouser/SelectRegularRolesSOUserTest.html">SelectRegularRolesSOUserTest</a></td></tr>
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/editsouserpassword/EditSOUserPasswordTest.html">EditSOUserPasswordTest</a></td></tr>
 <tr><td><a href="AddUserSOWebsiteTest.html">AddUserSOWebsiteTest</a></td></tr>
+<tr><td><a href="ViewProfileWebsiteTest.html">ViewProfileWebsiteTest</a></td></tr>
 <tr><td><a href="../../../../portalweb/socialoffice/users/user/addsouser/TearDownSOUserTest.html">TearDownSOUserTest</a></td></tr>
 </tbody></table>
 </body>
