--- portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/viewprofilewebsite/ViewProfileWebsiteTests.html
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
-<head>
-  <meta content="text/html; charset=UTF-8" http-equiv="content-type" />
-  <title>ViewProfileWebsiteTests</title>
-</head>
-<body>
-<table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
-<tr><td><b>ViewProfileWebsiteTests</b></td></tr>
-<tr><td><a href="../../../../portalweb/socialoffice/users/user/addsouser/AddSOUserTest.html">AddSOUserTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/socialoffice/users/user/selectregularrolessouser/SelectRegularRolesSOUserTest.html">SelectRegularRolesSOUserTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/socialoffice/users/user/editsouserpassword/EditSOUserPasswordTest.html">EditSOUserPasswordTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/socialofficeprofile/profile/addusersowebsite/AddUserSOWebsiteTest.html">AddUserSOWebsiteTest</a></td></tr>
-<tr><td><a href="ViewProfileWebsiteTest.html">ViewProfileWebsiteTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/socialoffice/users/user/addsouser/TearDownSOUserTest.html">TearDownSOUserTest</a></td></tr>
-</tbody></table>
-</body>
-</html>
