--- portal-web/test/functional/com/liferay/portalweb/socialofficeprofile/profile/viewprofilewebsite/ViewProfileWebsiteTests.java
@@ -1,42 +1,0 @@
-/**
- * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portalweb.socialofficeprofile.profile.viewprofilewebsite;
-
-import com.liferay.portalweb.portal.BaseTestSuite;
-import com.liferay.portalweb.socialoffice.users.user.addsouser.AddSOUserTest;
-import com.liferay.portalweb.socialoffice.users.user.addsouser.TearDownSOUserTest;
-import com.liferay.portalweb.socialoffice.users.user.editsouserpassword.EditSOUserPasswordTest;
-import com.liferay.portalweb.socialoffice.users.user.selectregularrolessouser.SelectRegularRolesSOUserTest;
-import com.liferay.portalweb.socialofficeprofile.profile.addusersowebsite.AddUserSOWebsiteTest;
-
-import junit.framework.Test;
-import junit.framework.TestSuite;
-
-/**
- * @author Brian Wing Shun Chan
- */
-public class ViewProfileWebsiteTests extends BaseTestSuite {
-	public static Test suite() {
-		TestSuite testSuite = new TestSuite();
-		testSuite.addTestSuite(AddSOUserTest.class);
-		testSuite.addTestSuite(SelectRegularRolesSOUserTest.class);
-		testSuite.addTestSuite(EditSOUserPasswordTest.class);
-		testSuite.addTestSuite(AddUserSOWebsiteTest.class);
-		testSuite.addTestSuite(ViewProfileWebsiteTest.class);
-		testSuite.addTestSuite(TearDownSOUserTest.class);
-
-		return testSuite;
-	}
-}
